first draft
Late better than never! @theo-armour I started this repository as the
place holder for default guidelines and will add a specific one to each
repository with a link to this repo and the files inside. I also added a
guideline page for Python projects. It might be a good idea to add one
for JavaScrip projects so we can all follow your style.